refactor(any2mochi): move Erlang converter
