+++ tools/any2mochi/x/erlang/README.md
@@ -1,0 +1,4 @@
+# Erlang Converter
+
+This package provides helpers that convert Erlang source files to Mochi.
+Conversion is done via a lightweight parser implemented in `parser/parser.escript`.
