+++ tools/any2mochi/x/erlang/convert.go
@@ -1,4 +1,4 @@
-package any2mochi
+package erlang
 
 import (
 	"fmt"
@@ -6,9 +6,9 @@ import (
 	"strings"
 )
 
-// ConvertErlang converts erlang source code to Mochi using the language server.
-func ConvertErlang(src string) ([]byte, error) {
-	funcs, err := parseErlangAST(src)
+// Convert converts Erlang source code to Mochi using the language server.
+func Convert(src string) ([]byte, error) {
+	funcs, err := parseAST(src)
 	if err != nil {
 		return nil, err
 	}
@@ -56,11 +56,22 @@ func ConvertErlang(src string) ([]byte, error) {
 	return []byte(out.String()), nil
 }
 
-// ConvertErlangFile reads the erlang file and converts it to Mochi.
-func ConvertErlangFile(path string) ([]byte, error) {
+// ConvertFile reads the Erlang file and converts it to Mochi.
+func ConvertFile(path string) ([]byte, error) {
 	data, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
-	return ConvertErlang(string(data))
+	return Convert(string(data))
+}
+
+func numberedSnippet(src string) string {
+	lines := strings.Split(src, "\n")
+	if len(lines) > 10 {
+		lines = lines[:10]
+	}
+	for i, l := range lines {
+		lines[i] = fmt.Sprintf("%3d: %s", i+1, l)
+	}
+	return strings.Join(lines, "\n")
 }
