+++ tools/any2mochi/x/erlang/parser/parser.escript
